Move the security manager under the FAB auth manager
