+++ airflow/api_connexion/endpoints/role_and_permission_endpoint.py
@@ -34,10 +34,10 @@ from airflow.api_connexion.schemas.role_and_permission_schema import (
     role_schema,
 )
 from airflow.api_connexion.types import APIResponse, UpdateMask
+from airflow.auth.managers.fab.fab_security.airflow_manager import AirflowSecurityManager
 from airflow.security import permissions
 from airflow.utils.airflow_flask_app import get_airflow_app
 from airflow.www.fab_security.sqla.models import Action, Role
-from airflow.www.security import AirflowSecurityManager
 
 
 def _check_action_and_resource(sm: AirflowSecurityManager, perms: list[tuple[str, str]]) -> None:
