+++ airflow/auth/managers/fab/fab_security/__init__.py
@@ -1,3 +1,4 @@
+#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -14,12 +15,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from __future__ import annotations
-
-from airflow.auth.managers.fab.fab_security.airflow_manager import AirflowSecurityManager
-
-
-class MockSecurityManager(AirflowSecurityManager):
-    VIEWER_VMS = {
-        "Airflow",
-    }
