+++ airflow/config_templates/default_webserver_config.py
@@ -20,7 +20,7 @@ from __future__ import annotations
 
 import os
 
-from airflow.www.fab_security.manager import AUTH_DB
+from airflow.auth.managers.fab.fab_security.base_manager import AUTH_DB
 
 # from airflow.www.fab_security.manager import AUTH_LDAP
 # from airflow.www.fab_security.manager import AUTH_OAUTH
