+++ tests/models/test_dagbag.py
@@ -38,13 +38,13 @@ from sqlalchemy.exc import OperationalError
 
 import airflow.example_dags
 from airflow import models, settings
+from airflow.auth.managers.fab.fab_security.airflow_manager import ApplessAirflowSecurityManager
 from airflow.exceptions import SerializationError
 from airflow.models import DagBag, DagModel
 from airflow.models.serialized_dag import SerializedDagModel
 from airflow.serialization.serialized_objects import SerializedDAG
 from airflow.utils.dates import timezone as tz
 from airflow.utils.session import create_session
-from airflow.www.security import ApplessAirflowSecurityManager
 from tests import cluster_policies
 from tests.models import TEST_DAGS_FOLDER
 from tests.test_utils import db
