+++ tests/test_utils/api_connexion_utils.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 from contextlib import contextmanager
 
 from airflow.api_connexion.exceptions import EXCEPTIONS_LINK_MAP
-from airflow.www.security import EXISTING_ROLES
+from airflow.auth.managers.fab.fab_security.airflow_manager import EXISTING_ROLES
 
 
 @contextmanager
